flight-data-scraping/flight_plan_tracking/requirements.txt
requests 
mysql-connector
python-dotenv
xmltodict